ArraySearches.java
This program will use the fileIO class to search through a file and create it into an array so the main class can search through it depending on the answer InsertionSort, SelectionSort and BubbleSort. Each of these are implemented in their own method and will out put the result to the screen.